+++ src/index.ts
@@ -5,11 +5,13 @@ import { add, Command } from './commands/add'
 import { reason } from './commands/reason'
 import { init } from './commands/init'
 import { snippets, SnippetLanguage, SnippetIDE } from './commands/snippets'
+import { graphql } from './commands/graphql'
 
 export interface CLIFlags {
   javascript: boolean
   ide?: string
   language?: string
+  examples?: boolean
 }
 
 export interface CLIProps {
@@ -42,6 +44,9 @@ export const run = (cli: meow.Result) => {
         ide: flags.ide as SnippetIDE,
       })
       break
+    case 'graphql':
+      graphql({ name, flags: flags as CLIFlags })
+      break
     default:
       console.log(help)
   }
@@ -58,11 +63,13 @@ const cli = meow(
     $ react <name> [flags]      Creates a React app (CRA)
     $ reason <name>             Creates a ReasonReact app
     $ snippets [flags]          Copy snippets to clipboard
+    $ graphql <name>            Create a GraphQL API
 
     Flags
     --javascript    JavaScript app (react)
     --ide           IDE for snippets (snippets) 
     --language      Language for snippets (snippets) 
+    --examples      GraphQL examples (examples)
     `,
   {
     flags: {
@@ -75,6 +82,9 @@ const cli = meow(
       },
       language: {
         type: 'string',
+      },
+      examples: {
+        type: 'boolean',
       },
     },
   }